# backend/main.go
package main

import (
	"context"
	"log"
	//"fmt"

	"github.com/beam19857/app/controllers"
	_ "github.com/beam19857/app/docs"
	"github.com/beam19857/app/ent"
	"github.com/gin-contrib/cors"
	"github.com/gin-gonic/gin"
	_ "github.com/mattn/go-sqlite3"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)
type Users struct {
		User []User
}
type User struct {
	DoctorName string
	DoctorEmail string
}
type Departments struct {
	Departmet []Depatment
}
type Depatment struct {
	DeparmentName string
}
type Expertises struct {
	Expertise []Expertise
}
type Espertise struct {
	ExpertiseName string
}
type Positions struct {
	Position []Position
}
type Position struct {
	PositionName string
}
// @title SUT SA Example API
// @version 1.0
// @description This is a sample server for SUT SE 2563
// @termsOfService http://swagger.io/terms/

// @contact.name API Support
// @contact.url http://www.swagger.io/support
// @contact.email support@swagger.io
// @license.name Apache 2.0
// @license.url http://www.apache.org/licenses/LICENSE-2.0.html

// @host localhost:8080
// @BasePath /api/v1

// @securityDefinitions.basic BasicAuth

// @securityDefinitions.apikey ApiKeyAuth
// @in header
// @name Authorization

// @securitydefinitions.oauth2.application OAuth2Application
// @tokenUrl https://example.com/oauth/token
// @scope.write Grants write access
// @scope.admin Grants read and write access to administrative information

// @securitydefinitions.oauth2.implicit OAuth2Implicit
// @authorizationUrl https://example.com/oauth/authorize
// @scope.write Grants write access
// @scope.admin Grants read and write access to administrative information

// @securitydefinitions.oauth2.password OAuth2Password
// @tokenUrl https://example.com/oauth/token
// @scope.read Grants read access
// @scope.write Grants write access
// @scope.admin Grants read and write access to administrative information

// @securitydefinitions.oauth2.accessCode OAuth2AccessCode
// @tokenUrl https://example.com/oauth/token
// @authorizationUrl https://example.com/oauth/authorize
// @scope.admin Grants read and write access to administrative information
func main() {
	router := gin.Default()
	router.Use(cors.Default())

	client, err := ent.Open("sqlite3", "file:ent.db?cache=shared&_fk=1")
	if err != nil {
		log.Fatalf("fail to open sqlite3: %v", err)
	}
	defer client.Close()

	if err := client.Schema.Create(context.Background()); err != nil {
		log.Fatalf("failed creating schema resources: %v", err)
	}

	v1 := router.Group("/api/v1")
	controllers.NewUserController(v1, client)
	controllers.NewDepartmentController(v1, client)
	controllers.NewExpertiseController(v1, client)
	controllers.NewPositionController(v1, client)

	// Set Department Data
	department := []string{"Antenatal","Operating Room","Deceased Receive","Treatment"}
	for _, r := range department {
		client.Department.
			Create().
			SetDepartmentName(r).
			Save(context.Background())
	}
		// Set Position Data
	position := []string{"Nurse","Doctor","Parcel worker","Assistant","Medical records officer"}
	for _, r := range position {
		client.Position.
			Create().
			SetPositionName(r).
			Save(context.Background())
	}


	router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	router.Run()

	
}